pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Multimodal MM-IMDB: early/late fusion, semi-supervised, consistency training"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.12.0",
    "kagglehub>=1.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
